create select season component
